creditrisk_rag_system/requirements.txt
streamlit
pandas
numpy
scikit-learn
matplotlib
seaborn
xgboost
langgraph
langchain-core
faiss-cpu
groq
python-dotenv
